requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
boto3==1.28.85
botocore==1.31.85
celery==5.3.6
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.2
django-celery-beat==2.7.0
django-environ==0.10.0
django-redis==5.3.0
django-timezone-field==7.0
djangorestframework==3.14.0
drf-yasg==1.21.8
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.4.2
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pytest==7.4.4
pytest-django==4.5.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==4.6.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
